+++ CHANGELOG.md
@@ -11,10 +11,48 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 - Python.Net 3.0 Support
 
-### Fixxed
+### Fixed
 
 - Converter bugfix for wrong value convert
 - Tree module refactoring
+
+### Changed
+
+- Keyword behaviour changed for Element Should Exist and Element Should Not Exist
+  - Keywords contains now a use_exceptions flag to decide if an exception should be called or a return value.
+    - Exception handling will be used in general to check if a ui element is closing or opened
+    - If you want to check it by your own use_exception flag should be set to ${FALSE}
+      - Wait Until Keyword Succeeds combination does not work anymore to check if an ui is displaying after amount of time
+      - Returns now always True or False
+
+#### Old Syntax
+```
+# This throws a flaui exception and stop test exeuction
+Element Should Exist  /WRONG/XPATH
+```
+
+#### New Syntax
+
+- You can now decide to use the old syntax or new style by using the flag parameter use_exceptions.
+- This is by default ${TRUE} to avoid this breaks for checkups if a ui element is displaying delayed.
+
+```
+# This throws a flaui exception and stop test exeuction
+Element Should Exist      /WRONG/XPATH
+Element Should Not Exist  /VALID_XPATH
+
+# Usage with Wait Until Keyword Succeeds
+Wait Until Keyword Succeeds  Element Should Exist      /VALID_XPATH
+Wait Until Keyword Succeeds  Element Should Not Exist  /VALID_XPATH
+
+# If you want to check the return value by your own syntax.
+# Wait Until Keyword Succeeds will not work anymore because this keyword will always Return True or False now
+${RESULT}  Element Should Exist  /VALID_XPATH  ${FALSE}
+${RESULT}  Element Should Not Exist  /WRONG/XPATH  ${FALSE}
+```
+
+- Wait Until Element Is Hidden does not check anymore if element exists
+  - No xpath not found exception is called anymore to avoid break tests
 
 ## [Release][1.6.6] [1.6.6][1.6.5-1.6.6] - 2021-09-01